Create README.md
This is a simple web-based Image Search Engine that allows users to search for images using keywords. It utilizes the Unsplash API to fetch and display images based on user queries. Additionally, the project includes a "Night mode" feature that allows users to toggle between light and dark themes for the interface.